@@ -1,9 +1,8 @@
-﻿namespace Challenge1.Models
+﻿namespace Challenge1.Models;
+
+public class Customer
 {
-	public class Customer
-	{
-		public int Id { get; set; }
-		public string FirstName { get; set; } = string.Empty;
-		public string LastName { get; set; } = string.Empty;
-	}
+	public int Id { get; set; }
+	public string FirstName { get; set; } = string.Empty;
+	public string LastName { get; set; } = string.Empty;
 }
